requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.18.1
networkx==3.4.2
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
patsy==1.0.2
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
squarify==0.4.4
statsmodels==0.14.6
streamlit==1.55.0
sympy
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vnai==2.4.6
vnstock==3.5.0
vnstock_ezchart==0.0.3
watchdog==6.0.0
wordcloud==1.9.6
zipp==3.23.0
